pyproject.toml
[tool.poetry]
name = "programmingfabot"
version = "0.1.0"
description = ""
authors = ["Abrorjon Ruziboev <abror.ruzibayev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.25.0"
alembic = "^1.9.1"
sqlalchemy = "^1.4.45"
python-telegram-bot = {version = "^20.0b0", allow-prereleases = true}
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
